move extras and create cache
