+++ sharkfuser/include/fusilli/support/asm_emitter.h
@@ -30,9 +30,9 @@
 
 #include "fusilli/attributes/tensor_attributes.h"
 #include "fusilli/attributes/types.h"
-#include "fusilli/extras.h"
 #include "fusilli/graph/graph.h"
 #include "fusilli/node/conv_node.h"
+#include "fusilli/support/extras.h"
 
 #include <cassert>
 #include <format>
